fix: Structural recursion in the presence of parentheses
This didn't work before
```
def f (n : Nat) : Nat :=
  match n with
  | 0 => 0
  | n + 1 => (f) n
```
because the `RecApp` metadata marker gets in the way. More practical,
such code seems to be produced when using `rw` or `simp` in recursive
theorems.

This creates variants of the `Expr.app`-decomposing functions that look
through that annotation and uses it in a few neuralgic points in the
structural recursion code.

Test case is included. Fixes #2810.