feat: `pp.beta` to apply beta reduction when pretty printing
This was a Lean 3 pretty printer option. While this pretty printer option tends to lead to confusing situations when set, it has been frequently requested. [It is possible](https://github.com/leanprover-community/mathlib4/pull/7910) to implement this pretty printer option as a user, but it comes with some artifacts (for instances, hovers not necessarily beta reduced). Adding this as a core pp option is cleanest.

(We should consider having hooks into the tactic evaluator to allow users to transform the tactic state between tactics. This would enable people to beta reduce the entire local context for real, which would be useful for teaching.)