feat: eliminators for `induction` and `cases` tactics
Defines `@[induction_eliminator]` and `@[cases_eliminator]` for defining custom eliminators for the `induction` and `cases` tactics, respectively.

Also adds `Nat.recAux` and `Nat.casesAuxOn`, which are eliminators that are defeq to `Nat.rec` and `Nat.casesOn` but use `0` and `n + 1`.
These will be given the above attributes in a followup commit after updating stage0.