feat: `pp.parens` option to pretty print with all parentheses
This PR adds the option `pp.parens` (default: false) that causes the pretty printer to eagerly parentheses, which can be useful for teaching and for understanding the structure of expressions.

Any notations with precedence greater than or equal to `maxPrec` do not receive such discretionary parentheses; this precedence level is considered to be infinity.

This option was a feature in the Lean 3 community edition.