refactor: WF.Eqns: remove unreachable fix-folding
I was about to to address the TODO

     /- TODO: check arity of the given function. If it takes a PSigma as the last argument,
        this function will produce incorrect results. -/

because we now have an arity-observing variant of `decodePackedArg?` in
`unpackArg` in `PackMutual`, and it would be prudent to use it here.

But I first wanted to create a test case that would actually exhibit
this corner case, and failed.

This code was added in 096e4eb6d0f and it had a test case, but not even
that test case seems to be actually using the `decodePackedArg?`
function (any more).

So this seems to be dead code, possibly due to other changes to the
system, and thus can be removed.

* [ ] Check mathlib if the above observation is correct
* [ ] If mathlib does not show a use of this feature, ask the experts to
  come up with a test case exercising it.
* [ ] Maybe: Bisect lean to find out why this is no longer needed for
  the original test case.